feat: 409 상태코드

+++ fqs.create.ts
@@ -9,5 +9,8 @@ export default function create(table_name: string, schema: Schema): void {
     const fields = file.fields;
 
     const save_csv = Papa.unparse({ fields });
+    if (fs.existsSync(table_name + ".csv")) {
+        throw { code: "EEXIST" };
+    }
     fs.writeFileSync(table_name + ".csv", save_csv);
 }
